have been creating to saving checkbox from TnlaAndCableCategory

+++ src/org/itstep/vinokurov/domain/TnlaAndCableCategory.java
@@ -3,22 +3,22 @@ package org.itstep.vinokurov.domain;
 import java.util.Set;
 
 public class TnlaAndCableCategory extends Entity{
-	private Short idTnla;
-	private Set<Short> idesOfCableCategory;
+	private Long idTnla;
+	private Set<Long> idesOfCableCategory;
 	
-	public Short getIdTnla() {
+	public Long getIdTnla() {
 		return idTnla;
 	}
 	
-	public void setIdTnla(Short idTnla) {
+	public void setIdTnla(Long idTnla) {
 		this.idTnla = idTnla;
 	}
 	
-	public Set<Short> getIdesOfCableCategory() {
+	public Set<Long> getIdesOfCableCategory() {
 		return idesOfCableCategory;
 	}
 	
-	public void setIdesOfCableCategory(Set<Short> idesCableCategory) {
+	public void setIdesOfCableCategory(Set<Long> idesCableCategory) {
 		this.idesOfCableCategory = idesCableCategory;
 	}
